Add deserializer for CreateActionArgs. Move serialization into it's own package.
